+++ src/main/java/com/parkify/vehitrack/controller/VisitorController.java
@@ -1,14 +1,12 @@
 package com.parkify.vehitrack.controller;
 
+import com.parkify.vehitrack.dto.VisitorDTO;
 import com.parkify.vehitrack.entity.Visitors;
 import com.parkify.vehitrack.service.VisitorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/visitors")
@@ -21,6 +19,14 @@ public class VisitorController {
     public ResponseEntity<Visitors> createVisitor(@RequestBody Visitors visitors) {
         Visitors visitor = visitorService.addVisitor(visitors);
         return new ResponseEntity<>(visitor, HttpStatus.OK);
+    }
+
+
+    @GetMapping("/getByRegistrationNumber")
+    public ResponseEntity<VisitorDTO> getVisitorByRegistrationNumber(
+            @RequestParam String registrationNumber) {
+        VisitorDTO visitorDTO = visitorService.getVisitorByRegistrationNumber(registrationNumber);
+        return new ResponseEntity<>(visitorDTO, HttpStatus.OK);
     }
 
 }